Split up prefixdb files
